The Spatial Reference field is no longer disabled when drawing plots, this allows the user to change the spatial reference field as if they were clicking on the map.
This was quite difficult change to get right, particularly avoiding breaking existing code. I have tested it in a few situations which applies to existing code, and also examined the code by eye so I hope it is OK, but another pair of eyes giving it a once over wouldn't do any harm.